#4893 Extract async inventory skeleton loader
