+++ indra/newview/CMakeLists.txt
@@ -105,6 +105,7 @@ set(viewer_SOURCE_FILES
     llappearancemgr.cpp
     llappviewer.cpp
     llappviewerlistener.cpp
+    llasyncinventoryskeletonloader.cpp
     llattachmentsmgr.cpp
     llaudiosourcevo.cpp
     llautoreplace.cpp
@@ -780,6 +781,7 @@ set(viewer_HEADER_FILES
     llappearancemgr.h
     llappviewer.h
     llappviewerlistener.h
+    llasyncinventoryskeletonloader.h
     llattachmentsmgr.h
     llaudiosourcevo.h
     llautoreplace.h
